+++ common/trinary/trit_long.h
@@ -4,13 +4,13 @@
 #include "common/stdint.h"
 #include "common/trinary/trits.h"
 
-long long trits_to_long(trit_t *trit, size_t const i);
-size_t min_trits(long long value);
-size_t long_to_trits(long long value, trit_t *trits);
+int64_t trits_to_long(trit_t *trit, size_t const i);
+size_t min_trits(int64_t value);
+size_t long_to_trits(int64_t value, trit_t *trits);
 
 size_t nearest_greater_power_of_three(size_t value, size_t *power);
-size_t encoded_length(long long value);
-int encode_long(long long value, trit_t *out, size_t size);
-long long get_encoded_long(trit_t *trits, size_t length, size_t *end);
+size_t encoded_length(int64_t value);
+int encode_long(int64_t value, trit_t *out, size_t size);
+int64_t get_encoded_long(trit_t *trits, size_t length, size_t *end);
 
 #endif
